fix for self a managing rules
moved set_rdescript to its own file, and added a call inside the merge rule constructor
