+++ castervoice/lib/ctrl/mgr/grammar_manager.py
@@ -1,5 +1,4 @@
 import os, traceback
-import re
 
 from dragonfly import Grammar
 
@@ -8,6 +7,7 @@ from castervoice.lib.ctrl.mgr.errors.invalid_companion_configuration_error impor
 from castervoice.lib.ctrl.mgr.errors.not_a_module import NotAModuleError
 from castervoice.lib.ctrl.mgr.loading.load.content_type import ContentType
 from castervoice.lib.ctrl.mgr.managed_rule import ManagedRule
+from castervoice.lib.ctrl.mgr.rule_formatter import set_rdescript
 from castervoice.lib.ctrl.mgr.rules_enabled_diff import RulesEnabledDiff
 from castervoice.lib.merge.ccrmerging2.hooks.events.activation_event import RuleActivationEvent
 from castervoice.lib.merge.ccrmerging2.sorting.config_ruleset_sorter import ConfigBasedRuleSetSorter
@@ -120,7 +120,7 @@ class GrammarManager(object):
             printer.out(invalidation)
             return
 
-        self._set_rdescript(rule_class.mapping, class_name)
+        set_rdescript(rule_class.mapping, class_name)
         '''
         rule should be safe for loading at this point: register it
         but do not load here -- this method only registers
@@ -377,25 +377,3 @@ class GrammarManager(object):
             GrammarManager._get_next_id.id = 0
         GrammarManager._get_next_id.id += 1
         return str(GrammarManager._get_next_id.id)
-
-    @staticmethod
-    def _set_rdescript(mapping, rcn):
-        for spec, action in mapping.items():
-            # pylint: disable=no-member
-            if hasattr(action, "rdescript") and action.rdescript is None:
-                mapping[spec].rdescript = GrammarManager._create_rdescript(spec, rcn)
-            elif hasattr(action, "rdescript") and not action.rdescript.startswith(rcn):                    
-                mapping[spec].rdescript = "%s: %s" % (rcn, action.rdescript)
-                
-
-    @staticmethod
-    def _create_rdescript(spec, rcn):
-        rule_name = rcn
-        for unnecessary in ["Non", "Rule", "Ccr", "CCR"]:
-            rule_name = rule_name.replace(unnecessary, "")
-        extras = ""
-        named_extras = re.findall(r"<(.*?)>", spec)
-        if named_extras:
-            extras = ", %(" + ")s, %(".join(named_extras) + ")s"
-        return "%s: %s%s" % (rule_name, spec, extras)
-
