+++ castervoice/lib/merge/mergerule.py
@@ -1,8 +1,10 @@
 import collections
-
-from dragonfly import MappingRule, Function
 import re
-from castervoice.lib import printer, available_commands_tracker
+
+from dragonfly import Function, MappingRule
+
+from castervoice.lib import available_commands_tracker, printer
+from castervoice.lib.ctrl.mgr.rule_formatter import set_rdescript
 from castervoice.lib.merge.ccrmerging2.pronounceable import Pronounceable
 
 
@@ -22,6 +24,7 @@ class MergeRule(MappingRule, Pronounceable):
         _mapping = mapping or self.mapping.copy()
         _extras = extras or self.extras[:]
         _defaults = defaults or self.defaults.copy()
+        set_rdescript(_mapping, _name)
         #
         super(MergeRule, self).__init__(name=_name,
                                         mapping=_mapping,
